Collaboration Service was not working in Stella.
We found that the service was depending on results from CODAP that were not present through the current version of the DI-API.
Specifically, the service depended on notifications providing full case representations, where notifications only provided case IDs.
Added an additional asynchronous step to this process to fetch the full case, upon notification. Aligned some of the logic a little
better with the CODAP API.

Some changes were made in parallel on the CODAP side. So, this Data Interactive will not be functional against released CODAP
until the next CODAP release (0371).