Animate ball returning from slot even when hidden
In order to support showing a bit of a mixer ball peaking out from the
hidden cover, we animate the ball selection, even though we're not
animating the mixing. However, usually the ball only returns from the
slot due to the mixing animation.

This adds a small additional animation that just pulls the ball back
from the mixer slot when the mixer is animation selection, if the model
is "hidden."

[#150865833]