Add savable password-protection to model options
This adds a password field, which sets a password that is saved to the
model state.

The presence of the password indicates that the model is locked. The
absence of a password indicates it is unlocked. Unlocking a model clears
the password, allowing it to be set to a new password if desired.

The UI shows an plain text field if there is no password. If a password
is set, the UI always shows a 10-character password field, regardless
of the length of the password.

The password is saved as part of the model state, and loading in a
password locks the model.

The user could determine the password by looking at the network traffic
to see plain-text password in the model's JSON, or by requesting the
model JSON from CODAP, if such a thing is possible. However, the user
cannot tell the password from any global variables in the model itself.

[#151721570]