Improve mixer anim efficiency for models 50+ cases
This improves the efficiency of animating the randomization step for mixer
models with over 50 balls, by increasingly skipping balls to animate as we
get more and more cases.

At 50 cases we'll pick one ball to skip every step. By 200 cases, we'll
skip half the balls each step.

This makes models with < ~300 or so cases better (depending on the machine),
but further improvements are needed above that.