Monitor attribute names for changes
After we create attributes for the first time, request them all again
to get IDs. Then monitor notifications for attribute name changes, and
update our internal attribute names as necessary.

When we send items, convert the item attributes to the new names.

The first thing that would make this easier is if we could assign IDs
to the attributes on creation, instead of having to request them after
creating them to get the IDs. We can almost do this, but CODAP seems
sensitive to the numbers (sometimes odd failures are thrown, which are
fixed by having different attributes being higher or lower), and also we
can't be certain we aren't trying to use existing IDs. Nicer would be to
use a string ID, but this doesn't seem to be supported.

The second thing that would make this easier would be if we could use
the attribute ID the whole time, sending items with keys being the
attribute ID, instead of the new name the user has picked.