Look for attr ids if we're loading old interactive
Even if we're using a pre-existing dataset (because the user had
already used the Sampler before the update), look for attributes that
match the names we expect and save their ids.

If the user has already updated the attributes to be something else,
the user's state is a lost cause. Dragging in a brand-new Sampler would
fix it for them, because it would create a new data set.