Shorten connect mode label to mitigate weird Safari bug.
In Safari, if you clicked on the label, the left border would shift offscreen to the left, obscuring part of the plugin. Error could not be fixed by expanding the plugin, but it _could_ be fixed by clicking on the label again. This is a Safari-only issue and hopefully transient.